Renamed a lot of index.js files
