+++ face-analyzer/src/layout/MainLayout/Header/ProfileSection/ProfileSection.js
