+++ face-analyzer/src/layout/MainLayout/LogoSection/LogoSection.js
