+++ face-analyzer/src/layout/MainLayout/Sidebar/MenuList/NavItem/NavItem.js
