+++ face-analyzer/src/layout/MainLayout/Sidebar/Sidebar.js
@@ -9,8 +9,8 @@ import PerfectScrollbar from 'react-perfect-scrollbar';
 import { BrowserView, MobileView } from 'react-device-detect';
 
 // project imports
-import MenuList from './MenuList';
-import LogoSection from '../LogoSection';
+import MenuList from './MenuList/MenuList';
+import LogoSection from '../LogoSection/LogoSection';
 import { drawerWidth } from 'store/constant';
 
 // ==============================|| SIDEBAR DRAWER ||============================== //
