+++ face-analyzer/src/views/stimuli/Stimuli.js
