+++ face-analyzer/src/views/stimuli/charts/Stats.js
